@@ -52,13 +52,13 @@
     "query-string": "^6.2.0",
     "raf": "3.4.0",
     "rc-progress": "^2.2.5",
-    "react": "^16.6.0",
+    "react": "^16.7.0-alpha.0",
     "react-ace": "^6.1.4",
     "react-autosize-textarea": "^4.0.0",
     "react-dev-utils": "^5.0.2",
     "react-dnd": "^5.0.0",
     "react-dnd-html5-backend": "^5.0.1",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.7.0-alpha.0",
     "react-icons": "^3.1.0",
     "react-notifications": "^1.4.3",
     "react-redux": "^5.0.7",
