@@ -1,101 +1,93 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { FlexContainer, Paper } from '@edulastic/common';
+import React, { useState, useEffect, memo } from 'react';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import { withRouter } from 'react-router-dom';
 
-import ItemHeader from '../QuestionEditor/ItemHeader';
-import TestPageNav from './TestPageNav';
 import { Container } from '../common';
-import Items from './Items';
-import { getTestItemsSelector, getItemsLoadingSelector } from '../../selectors/testItems';
-import { receiveTestItemsAction } from '../../actions/testItems';
+import AddItems from './AddItems';
+import TestPageHeader from './TestPageHeader';
+import {
+  createTestAction,
+  receiveTestByIdAction,
+  setTestDataAction,
+  updateTestAction,
+} from '../../actions/tests';
+import { getTestSelector } from '../../selectors/tests';
 
-class TestPage extends Component {
-  state = {
-    current: 'addItems',
-    checkedItems: [],
+const TestPage = ({ createTest, match, receiveTestById, test, setData, updateTest }) => {
+  useEffect(() => {
+    if (match.params.id) {
+      receiveTestById(match.params.id);
+    }
+  }, []);
+
+  const [current, setCurrent] = useState('addItems');
+
+  const handleNavChange = (value) => {
+    setCurrent(value);
   };
 
-  static propTypes = {
-    match: PropTypes.object.isRequired,
-    testItems: PropTypes.array.isRequired,
-    receiveTestItems: PropTypes.func.isRequired,
-    loading: PropTypes.bool.isRequired,
+  const handleAddItems = (items) => {
+    setData({ ...test, testItems: items.map(id => ({ id })) });
   };
 
-  componentDidMount() {
-    const { receiveTestItems } = this.props;
-    receiveTestItems();
-  }
-
-  handleNavChange = (value) => {
-    this.setState({
-      current: value,
-    });
-  };
-
-  handleSelectItems = (checkedItems) => {
-    this.setState({
-      checkedItems,
-    });
-  };
-
-  handleAddItems = () => {
-    console.log('handleAddItems');
-  };
-
-  renderContent = () => {
-    const { testItems, loading } = this.props;
-    const { current, checkedItems } = this.state;
-
+  const renderContent = () => {
     switch (current) {
       case 'addItems':
-        return (
-          <Items
-            items={testItems}
-            loading={loading}
-            onSelect={this.handleSelectItems}
-            checkedItems={checkedItems}
-            onAddItems={this.handleAddItems}
-          />
-        );
+        return <AddItems onAddItems={handleAddItems} />;
       default:
         return null;
     }
   };
 
-  render() {
-    const { match } = this.props;
-    const { current } = this.state;
+  const handleSave = () => {
+    if (test.id) {
+      updateTest(test.id, test);
+    } else {
+      createTest(test);
+    }
+  };
 
-    return (
-      <div>
-        <ItemHeader
-          hideIcon
-          title="Test"
-          link={{ url: '/author/tests', text: 'Test list' }}
-          reference={match.params.id || ''}
-        >
-          <FlexContainer>
-            <TestPageNav onChange={this.handleNavChange} current={current} />
-          </FlexContainer>
-        </ItemHeader>
-        <Container>
-          <Paper>{this.renderContent()}</Paper>
-        </Container>
-      </div>
-    );
-  }
-}
+  return (
+    <div>
+      <TestPageHeader
+        onChangeNav={handleNavChange}
+        current={current}
+        onSave={handleSave}
+        title={test.title}
+      />
+      <Container>{renderContent()}</Container>
+    </div>
+  );
+};
+
+TestPage.propTypes = {
+  createTest: PropTypes.func.isRequired,
+  updateTest: PropTypes.func.isRequired,
+  receiveTestById: PropTypes.func.isRequired,
+  setData: PropTypes.func.isRequired,
+  match: PropTypes.object.isRequired,
+  test: PropTypes.object,
+};
+
+TestPage.defaultProps = {
+  test: null,
+};
 
 const enhance = compose(
+  memo,
+  withRouter,
   connect(
     state => ({
-      testItems: getTestItemsSelector(state),
-      loading: getItemsLoadingSelector(state),
+      test: getTestSelector(state),
     }),
-    { receiveTestItems: receiveTestItemsAction },
+    {
+      createTest: createTestAction,
+      updateTest: updateTestAction,
+      receiveTestById: receiveTestByIdAction,
+      setData: setTestDataAction,
+    },
   ),
 );
 
